Introduce the `delegateCiphertext()` precompile
The `delegateCiphertext()` precompiled contract delegates a ciphertext
to callers one level below the stack depth at which a ciphertext is
verified. That allows contracts to return handles. See the
`Caller.reencrypt_with_delegate()` method in `handles.sol` for an
example.

Make sure SLOAD doesn't overwrite any already verified ciphertexts in
memory. That allows use of `delegateCiphertext()`.

In a future commit, implement `delegateCiphertext()` as part of the
RETURN opcode.

Refactor code for readability.

Add use cases in the sample `handles.sol` file that show automatic
delegation (deeper in the stack) and manual delegation to the caller
stack depth via `delegateCiphertext()`. Also, split the precompiled
contracts in a separate "library" `Precompiles` contract that is to be
inherited by user contracts.