Add support for automatic delegation on RETURN
On the RETURN opcode, check if any known ciphertext handle is part of
the returned buffer. If that's the case, consider it a delegation
request and make the handle available to the the caller.

Change the value of the `verifiedCiphertexts` map from a value to a
pointer. That aids in updating the map and avoids copying potentially
big ciphertexts.

Add examples in `handles.sol` for automatic delegation.