Skip FHE ops and requires if not committing
If results from the current EVM execution won't be committed:
1. Skip FHE execution by generating random ciphertexts in FHE
precompiles.
2. Don't evaluate require statements and assume they are true.

Above changes are useful when doing gas estimation that, in itself,
might be expensive and have side effects on the oracle DB (before the
transaction is actually added in a block).

Since reencrypt is only allowed on EthCall anyway, don't do anything
special for it in case not committing.