Do not import random ciphertext on EthCall
We want to import random ciphertexts only during simulation, e.g. gas
estimation. Therefore, if the `EthCall` flag is set, even though the
`Commit` flag is not set, we don't want to import a random ciphertext.
That allows for FHE operations in read-only calls (i.e. EthCall).

Add support for the FHE less-than (lt) operation as a precompiled
contract.