Add the `rand` precompiled contract (#44)
The `rand` precompiled contract returns a ciphertext that encrypts a
random value. It uses the chacha20 cipher for that. The cipher is seeded
on every call. The seed is determined by:
```
seed = sha3-256(globalSeed, contractAddress, perContractCounter)
```
We persist `perContractCounter` in the contract's protected storage and
increment it by 1 on every call. We increment it immediately after reading it.
Rationale is that if something fails or if the caller doesn't check for the
return value of the precompile (maliciously or not), we might end up
generating the same value multiple times.

The output from the cipher is trivially encrypted using tfhe-rs. Trivial
encryption is deterministic.

Currently, the seed inputs and the cipher output are public. Our
intention is to implement random generation in FHE, making them private.
We expose the `rand` precompiled contract without privacy till then,
such that contracts can be written with it and experiments can be made.