Add support for optimistic ciphertext requires
In order to reduce latency from communication with the decryption
oracle, add support for optimistic ciphertext requires. They work by
maintaining a product of all optimistic requires during execution. Since
require ciphertexts are either 0 or 1, by homomorphically multiplying
them all, we will get 1 at the end if all are 1 or 0 if at least 0 is 0.
That allows us to only decrypt a single value at the end via the
decryption oracle.

The downside of this approach is that users will pay the full gas amount
for the transaction as we need to fully execute it, assuming optimistic
require statements are true. Still, users can choose whether to use
them or not.

Finally, fix the return value of the existing ciphertext require to
`ErrExecutionReverted` such that some gas is refunded in case of a false
require.