Determine ciphertext size and msg modulus from keys
Instead of hardcoding values for ciphertext size and message modulus,
determine them from the FHE keys in use.

For ciphertext size, use trivial encryption at startup to encrypt a
a value and use the resulting ciphertext buffer size.

For message modulus, extract the value from the parameters that reside
in the client `cks` key.