Delegate handles found in call arguments
If a ciphertext handle is part of the arguments to a call opcode (call,
static, delegate, callCode), delegate it such that it is verified for
the callee. Previously, the delegation was implicit for deeper call
depths. This PR makes it explicit, vie either the `delegateCiphertext`
precompile or a call opcode.

Implement by keeping a set of verified stack depths per ciphertext.

Increment the stack depth when calling a precompile such that the depth
is correct inside the precompile itself.

Closes #52.