Delegate handles found in call arguments
If a ciphertext handle is part of the arguments to a call opcode (call,
static, delegate, callCode), delegate it such that it is verified for
the callee. Previously, the delegation was implicit for deeper call
depths.

Implement by keeping depth ranges (a pair of from and to depths) per
ciphertext. A ciphertext handle is verified at depth d if from <= d and
to == d.

Add support for a list of depth ranges per ciphertext. Rationale is that
we want to preserve verifications of ciphertexts in case a callee deeper
in the stack does not return a handle. Say handle h is verified at depth
1. Depth 2 can also verify it. However, if depth 2 does not return it,
we want h to still be verified at depth 1.