Add fractional CALLDATA gas cost
Since an FHE-based chain could require relatively big FHE ciphertexts to
be passed as transaction inputs, reduce the CALLDATA gas cost to
`1 / TxFheDataFractionalGas` per byte. Currently, it costs 1/4 gas
per byte. It doesn't matter if a byte is zero or non-zero, the cost is
the same. The `TxFheDataFractionalGas` constant could be changed in the
future as we gain more usage patterns from live networks.