Add gas costs for FHE operations
Add gas costs for FHE-related operations implemented both as precompiled
contracts and EVM opcodes. For operations whose runtime and/or storage
depend on FHE integer types, we define a base cost for the smallest
type and measure how much more resources bigger types require compared
to the base. That is how we determine costs for bigger types.

Base costs are somewhat arbitrary at that point. We would most likely
have to tweak them once we have parallel tfhe-rs operations and/or
hardware acceleration.

Costs for reencryption and requires are arbitrary at that point, because
we don't know how expensive they would be when done via threshold
protocols.

As of now, optimistic requires only support FheUint32. Reason is we
don't have tfhe-rs type casting such that we could have different types
for the control bit and the input values.

Currently, we don't handle refunds for SSTORE when persisting a
ciphertext handle. Also, we don't charge different costs for protected
storage based on whether a byte of ciphertext is zero or non-zero.

When doing SLOAD for a ciphertext that is already in memory, skip
storage and just import the same ciphertext at the current depth.

Finally, fix a bug in the `require` precompile that wasn't checking if
the input ciphertxt is verified - it was only checking it is an existing
one.