Add logging to the EVM
Add a logger to the EVM by introducing the `Logger` interface and a
default implementation that logs to stdout. The `Logger` interface is
compatible with the Tendermint/Cosmos one and, therefore, can be easily
hooked.

Add log statements to all FHE-related error cases. Currently, no debug
logs have been added, because the EVM might be called many times, e.g.
during gas estimation, etc. If it turns out we need them, we will add
them in a subsequent commit.

Comment out the `cast` precompile as we don't have an implementation for
it yet.