Only store verified ciphertexts to protected storage (#96)
On an SSTORE, first check if the value is a handle to a verified
ciphertext at the current depth. It is not enough to check if the
ciphertext is in protected memory.

Also, do not increment the depth when going inside a precompiled
contract. That leads to a ciphertext that is verified at one level
deeper than the correct one (after returning from the precompile).
Increment was added wrongly due to the now unsupported delegate
precompile. For example, if a a contract uses the add precompile to
produce a ciphertext at depth N, we want that a subsequent SSTORE at
the same depth N finds a verified ciphertext at N and not N + 1.

Fix unit tests that didn't set the EVM depth.