Do not delegate ciphertexts to precompiles on CALL
Rationale is that if a smart contract at depth N calls a precompile by a
CALL opcode (or its variants), there is no corresponding call to the
RETURN opcode. Therefore, the ciphertext will remain delegated to depth
N + 1 and that might not be the smart contract's original intention.

The reason is that if depth N + 1 is EVM bytecode, it will have to go
back to N via RETURN. However, if N + 1 is a precompile, it will return
"immediately" in Go.

To fix, do not delegate to precompiles on CALL opcode variants.