Add support for compact pub keys and ciphertexts
The public key `pks` is now a `CompactPublicKey` in tfhe-rs terminology.

For inputs, we use compact lists from tfhe-rs. Essentially, transaction
inputs are serialized tfhe-rs compact lists. In the verifyCiphertext
precompile, we deserialize the list and then expand it to a normal
ciphertext. From that point on, nothing changes and we compute on and
persist expanded ciphertexts. Ciphertexts in protected storage are not
compact and are always persisted in their expanded form.

Use trivial encryption to determine expanded ciphertext size at startup.

Remove secret key encryption in tfhe_test.go - instead, only use public
key encryption. The `pks` key is expected to always be available to the
node. We will use that later for serving it to clients (potentially
over a special precompile for that purpose).

Add tests for the verifyCiphertext precompile.

Note: As of now, we cannot detect arbitrary ciphertexts during
deserialization time. Therefore, verifyCiphertext might or might not
fail and might produce random ciphertexts. Will work on that separately.