Fix compact deserialization
When calling deserializeCompact(), we should not set the serialization
member to the input byte array, because the input is compact and the
serialization is expanded.

Remove "random" tfheCiphertexts and, instead, always use public key
encryption during gas estimation. Rationale is that this ensures valid
ciphertexts in all code paths and it also ensures the actual ciphertext
is random-looking bytes.

Always persists ciphertexts in protected storage during opSstore, even
if the Commit flag is not set. Reason is that we want the same code
paths during gas estimation and actual transaction. If we skip
persisting, opSload will behave differently on gas estimation and
transactions.

Do not skip ciphertext verification on verifyCiphertext when the Commit
flag is not set. Rationale is, again, that we want the same code path
for gas estimation and transactions.