Optimize gas estimation running time
Optimize gas estimation running time with two approaches:
1. Don't call any storage methods on SSTORE during gas estimation.
Instead, return early if we are in gas estimation (Commit = false). For
that to work, make sure that on SLOAD, we first lookup the ciphertext
hash in memory and, if found, use it without going to storage. That
improves performance by not doing heavy work of splitting a ciphertext
to 32 byte slots and persisting it to a memory-backed store for no
reason.
2. When importing a random ciphertext during gas estimation, do not use
public key encryption. Instead, just genearate an unique hash for the
random ciphertext by utilizing a counter that is part of the EVM itself.
That improves performance by not doing public key encryption and
serialization just to compute a ciphertext hash (handle).