Create protected storage before the actual contract
On the CREATE and CREATE2 opcodes, first create the protected storage
contract and then the actual contract. We want to do it in that order,
because, otherwise, saving ciphertexts to protected storage in Solidity
constructors doesn't work. Reason is that constructors run during the
creation of the actual contract, potentially persisting data to the
corresponding protected storage contract. However, if we call create()
for the protected storage contract afterwards, the writes in the
constructor would be lost. Creating the protected storage contract first
solves it.