Gas costs reflecting latest tfhe-rs benchmarks
To determine costs, we base calculations on the multiply operation as it
is the most expensive one - it takes around 0.5 seconds on a metal AWS
instance for FheUint32. We don't want computation to take more than 20
seconds, so we assume we can do ~30-40 multiplications in a block. With
block gas limit of 10M, we have 270K gas per multiply * ~35 = ~10M gas.
Gas costs of all other operations are computed as relative to FheUint32
multiply, based on their respective runtimes.

Finally, we make SLOAD and SSTORE cheaper as we assume the blockchain is
going to have a storage overhead and we don't want huge penalties for
users. That is subject to changes and different tokenomics.