Add support for explicit decryption
Implement as a `decrypt` precompile that takes a ciphertext handle and
returns a 32-byte array of a 256-bit big-endian unsigned integer.

Note that decrypt can be called in "oracle" mode only, i.e. on a
validator node that has the FHE secret key. Furthermore, we don't store
the decrypted result in the oracle/requires DB, making catching up of
full nodes that don't have the FHE secret key impossible. Reason is that
if decrypt is called in a txn, a full node will behave differently than
a validator node. We will work on that in a future commit.