Add support for explicit decryption
Implement as a `decrypt` precompile that takes a ciphertext handle and
returns a 32-byte array of a 256-bit big-endian unsigned integer.

For now, we assume all validators have the full FHE secret key.
Therefore, we start allowing encrypted require, optimistic require and
decrypt in view functions (i.e. over the EthCall RPC) executed on
validators. If executed on a full node, the view function would fail,
though, as it would try fetch the decryption from the DB and it won't be
there (unless the decryption was previously decrypted in a non-view
function). That implementation will change once we start using threshold
decryption.

Use bit AND for optimistic requires instead of 8 bit multiplication.
Rationale is that since we ensure ebool values are either 0 or 1, using
AND is faster than multiplication. Furthermore, don't do the AND when
encountering the optimistic require. Instead, accumulate requires and do
the AND operations at the end. That way, if a reversal happens for a
different reason, we wouldn't have executed ANDs and would save some
execution time.

If a decryption or an encrypted require is encountered when optimistic
requires have been used before, we decrypt the AND-ed optimistic
requires as part of the decryption or the decrypted require. Rationale
is that optimistic requires only revert at the end of the transaction.
However, we don't want to be decrypting (in require or decrypt) for no
reason. That limits the amount of data leaked in case of an optimistic
require reversal.

Add unit tests for encrypted require, optimistic require and decryption.