Fix gas estimation for explicit decryption
As of now, when doing gas estimation and we encounter an explicit
decryption request, we immediately return 32 zero bytes (i.e. a 0
uint256 value). However, that might be problematic, because gas costs in
different contexts depend on the value and amount of 0 bytes in it (e.g.
SSTORE or memory operations). A solution is to always return the
maximum possible decrypted value of 32 0xFF bytes, making gas estimation
more conservative, but making the txn much less likely to run out of
gas.

Thanks to @remi-gai for finding the issue.

Finally, check if the input ciphertext is verified on gas estimation
before returning the 32 0xFF bytes. That way if decrypt is called with
an unverified handle, we will fail early (during gas estimation).